+++ app/src/main/java/com/skymapglobal/cctest/workspace/newsfeed/domain/model/NewsResp.kt
@@ -5,7 +5,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class NewsResp(
-    val totalResults: Long,
+    val totalResults: Int,
     val articles: List<Article>?
 ) : Parcelable
